agrigrigri

+++ Projet-Fil-Rouge/Fil-Rouge-Backend/src/main/java/com/dev/filrouge/controller/ListeProduitsAdmin.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.dev.filrouge.model.Commande;
 import com.dev.filrouge.model.Produit;
+import com.dev.filrouge.model.ProduitCommande;
 import com.dev.filrouge.service.CommandeService;
 import com.dev.filrouge.service.ProduitService;
 
@@ -27,20 +28,20 @@ public class ListeProduitsAdmin {
 	
 	@DeleteMapping(value="/{id}")
     public void deleteProduit(@PathVariable Long id) {
-//    	boolean estPresent=false;
-//    	List<Commande> commandes = commandeService.commandeRepository.findAll();
-//    	FIN:for(Commande commande : commandes) {
-//    		List<Produit> produits = commande.getProduit();
-//    		for(Produit produit : produits) {
-//    			if(produit.getId().equals(id)) {
-//    				estPresent=true;
-//    				break FIN;
-//    			}
-//    		}
-//    	}
-//    	System.out.println(estPresent);
-//    	if(!estPresent) {
+    	boolean estPresent=false;
+    	List<Commande> commandes = commandeService.commandeRepository.findAll();
+    	FIN:for(Commande commande : commandes) {
+    		List<ProduitCommande> produits = commande.getProduitCommandes();
+    		for(ProduitCommande produit : produits) {
+    			if(produit.getId().getProduit().getId().equals(id)) {
+    				estPresent=true;
+    				break FIN;
+    			}
+    		}
+    	}
+    	System.out.println(estPresent);
+    	if(!estPresent) {
 			produitService.produitRepo.delete(produitService.produitRepo.findById(id).get());
-    	//}
+    	}
     }
 }
